+++ packages/soda-hooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-hooks
 
+## 6.5.19
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.23.0
+
 ## 6.5.18
 
 ### Patch Changes