+++ packages/soda-next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # soda-next
 
+## 0.4.33
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.23.0
+    - soda-hooks@6.5.19
+
 ## 0.4.32
 
 ### Patch Changes